refactor(userBootstrapping): Reduce API calls by adjusting components
